Don't shadow the standard gradle 'ext' field on tasks.
All Gradle Tasks (and many other things) have a special 'ext' extension,
which is of type ExtraPropertiesExtension
(see https://docs.gradle.org/current/javadoc/org/gradle/api/plugins/ExtensionAware.html).
Previously the DockerComposeExtension was shadowing this as it was set as a field
with name 'ext' on the gradle tasks created by this plugin.